+++ frontend/src/components/ui/TabPanel.tsx
@@ -1,0 +1,25 @@
+import { ReactNode } from "react";
+import { Box } from "@mui/material";
+
+type TabPanelProps = {
+  children: ReactNode;
+  index: number;
+  value: number;
+};
+
+export const TabPanel = (props: TabPanelProps) => {
+  const { children, value, index, ...other } = props;
+
+  return (
+    <div
+      role="tabpanel"
+      hidden={value !== index}
+      id={`tabpanel-${index}`}
+      aria-labelledby={`tab-${index}`}
+      // eslint-disable-next-line react/jsx-props-no-spreading
+      {...other}
+    >
+      {value === index && <Box sx={{ p: 3 }}>{children}</Box>}
+    </div>
+  );
+};
